Move UseSessions to sessions/http/
