+++ packages/core/src/sessions/http/use-sessions.hook.spec.ts
@@ -22,17 +22,17 @@ import {
   isHttpResponseRedirect,
   isHttpResponseUnauthorized,
   ServiceManager
-} from '../core';
+} from '../../core';
 import {
   SESSION_DEFAULT_ABSOLUTE_TIMEOUT,
   SESSION_DEFAULT_COOKIE_NAME,
   SESSION_DEFAULT_INACTIVITY_TIMEOUT,
   SESSION_USER_COOKIE_NAME
-} from './constants';
-import { readSession } from './core/read-session';
-import { Session } from './core/session';
-import { SessionState } from './core/session-state.interface';
-import { SessionStore } from './core/session-store';
+} from '../constants';
+import { readSession } from '../core/read-session';
+import { Session } from '../core/session';
+import { SessionState } from '../core/session-state.interface';
+import { SessionStore } from '../core/session-store';
 import { UseSessions } from './use-sessions.hook';
 
 describe('UseSessions', () => {
