+++ packages/core/src/sessions/index.ts
@@ -3,4 +3,4 @@ export { readSession } from './core/read-session';
 export { Store, SessionStore, SessionAlreadyExists } from './core/session-store';
 export { Session } from './core/session';
 export { SessionState } from './core/session-state.interface';
-export { UseSessions } from './use-sessions.hook';
+export { UseSessions } from './http/use-sessions.hook';
